% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_bycatch.R
\docType{data}
\name{metier_palette}
\alias{metier_palette}
\title{Colour palette for métier level 4 categories}
\format{
A named character vector.
}
\usage{
metier_palette
}
\description{
A named character vector mapping métier level 4 codes to hexadecimal colour
values used in bycatch plots.
}
\details{
The names correspond to métier level 4 codes and the values are hex colour
strings. This palette is used as the default fill scale in the interactive
bycatch and BPUE plotting functions.
}
\examples{
metier_palette["OTB"]

}
\keyword{datasets}
