% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vms.R
\name{plot_effort_map_app}
\alias{plot_effort_map_app}
\title{Plot effort layer with ecoregion outline and land. plot has data update date stamp}
\usage{
plot_effort_map_app(
  effort,
  ecoregion_name,
  ecoregion_shape,
  land_shape,
  fishing_category,
  crs,
  data_update_date,
  yr
)
}
\arguments{
\item{effort}{sf spatial object giving fishing effort}

\item{ecoregion_name}{character name of ecoregion}

\item{ecoregion_shape}{sf spatial object of ecoregion polygon}

\item{land_shape}{sf spatial object of land}

\item{fishing_category}{character vector given by user input. Should be a gear type or "all"}

\item{crs}{a valid crs string}

\item{data_update_date}{character -month and year or more specific to include in plot}

\item{yr}{numeric giving Year of vms data update}
}
\description{
Plot effort layer with ecoregion outline and land. plot has data update date stamp
}
