% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vms.R
\name{plot_sar_map_app}
\alias{plot_sar_map_app}
\title{Plot sar layer with ecoregion outline and land. plot has data update date stamp}
\usage{
plot_sar_map_app(
  sar_data,
  ecoregion_name,
  ecoregion_shape,
  land_shape,
  sar_layer,
  crs,
  data_update_date,
  yr
)
}
\arguments{
\item{sar_data}{sf spatial object giving benthic impact of fishing}

\item{ecoregion_name}{character name of ecoregion}

\item{ecoregion_shape}{sf spatial object of ecoregion polygon}

\item{land_shape}{sf spatial object of land}

\item{sar_layer}{character vector given by user input. Should be a "surface", "subsurface" or "all"}

\item{crs}{a valid crs string}

\item{data_update_date}{character -month and year or more specific to include in plot}

\item{yr}{numeric giving Year of vms data update}
}
\description{
Plot sar layer with ecoregion outline and land. plot has data update date stamp
}
