% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vms.R
\name{render_vms}
\alias{render_vms}
\title{Function to find and display vms images from www/vms folder}
\usage{
render_vms(ecoregion, gear, vms_layer, ns)
}
\arguments{
\item{ecoregion}{character}

\item{gear}{fishing gear, character, given by input$fishing_cat_selector}

\item{vms_layer}{character "sar" or "effort"}

\item{ns}{namespace from server definition}
}
\description{
Function to find and display vms images from www/vms folder
}
