% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\docType{data}
\name{revision_dates}
\alias{revision_dates}
\title{Revision dates for ecoregions
A named character vector mapping ecoregion acronyms to last revision dates of the FOs text.
 Used in the Overview tab.}
\format{
An object of class \code{character} of length 11.
}
\usage{
revision_dates
}
\description{
Revision dates for ecoregions
A named character vector mapping ecoregion acronyms to last revision dates of the FOs text.
 Used in the Overview tab.
}
\keyword{datasets}
