% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vms.R
\name{vms_bundle_content}
\alias{vms_bundle_content}
\title{Bundle vms content for download, either effort or sar data}
\usage{
vms_bundle_content(selected_ecoregion, vms_layer)
}
\arguments{
\item{selected_ecoregion}{reactive value}

\item{vms_layer}{character, either "sar" or "effort"}
}
\description{
Bundle vms content for download, either effort or sar data
}
