% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vms.R
\name{vms_bundle_filename}
\alias{vms_bundle_filename}
\title{Generate filename for vms download bundle}
\usage{
vms_bundle_filename(selected_ecoregion, vms_layer)
}
\arguments{
\item{selected_ecoregion}{}

\item{vms_layer}{}
}
\value{
character
}
\description{
Generate filename for vms download bundle
}
\examples{
vms_bundle_filename("Greater North Sea", "effort")
}
