% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vms.R
\name{vms_bundle_filename}
\alias{vms_bundle_filename}
\title{Generate filename for vms download bundle}
\usage{
vms_bundle_filename(selected_ecoregion, vms_layer)
}
\arguments{
\item{selected_ecoregion}{A reactive or function returning the selected ecoregion.}

\item{vms_layer}{Character string such as "effort" or "sar".}
}
\value{
A function that generates the zip filename.
}
\description{
Generate filename for vms download bundle
}
\examples{
selected_ecoregion <- function() "Greater North Sea"
vms_bundle_filename(selected_ecoregion, "effort")()
}
